requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
pydantic[email]==2.9.2
motor==3.6.0
python-dotenv==1.0.1
